refactor install script

+++ arch.sh
@@ -6,7 +6,6 @@ if (! which gcc || ! which pypy3 || ! which nc || ! which git) &>/dev/null; then
          gcc \
          git \
          man \
-         openbsd-netcat \
          pypy3 \
          python
 fi
@@ -26,19 +25,20 @@ cd /mnt
         git clone https://github.com/nathants/s4
     fi
     cd s4
-    # cli runs python, faster startup
-    sudo python3 -m pip install -r requirements.txt
+    if [ ! -f /tmp/requirements.done ]; then
+        sudo python3 -m pip install -r requirements.txt
+        sudo pypy3   -m pip install -r requirements.txt
+        touch /tmp/requirements.done
+    fi
     sudo python3 setup.py develop
-    # server runs pypy, faster in general
-    sudo pypy3 -m pip install -r requirements.txt
-    sudo pypy3 setup.py develop
+    sudo pypy3   setup.py develop
 )
 
 if ! which xxh3 &>/dev/null; then
     git clone https://github.com/nathants/bsv
     (
         cd bsv
-        make xxh3
-        sudo mv -fv bin/xxh3 /usr/local/bin
+        make
+        sudo mv -fv bin/* /usr/local/bin
     )
 fi
